requirements.txt
streamlit 
plotly==5.13.1
gunicorn
pandas
joblib
scikit-learn
gspread==5.12.2
datetime
python-dotenv
google-oauth2-tool==0.0.3
oauth2client
pytz